Massive chagnes to the intellij plugin.
Instead of calling the compiler and writing files, it now uses the augment
provider api. While this technically make it more britle to gradle plugin
changes, it is far more performant and prevents file change conflicts. Also,
because the whole gradle model isn't needed, it allows everything to be
availaible immedatly on start of the ide without refreshing.